+++ lib/dartapi_db.dart
@@ -2,11 +2,15 @@ library;
 
 export 'core/db_result.dart';
 export 'core/dartapi_db_core.dart';
+export 'core/db_transaction.dart';
 
 export 'drivers/postgres/postgres_database.dart';
 export 'drivers/mysql/mysql_database.dart';
+export 'drivers/sqlite/sqlite_database.dart';
 
 export 'types/db_config.dart';
 export 'types/db_type.dart';
 export 'types/pool_config.dart';
 export 'factory/database_factory.dart';
+
+export 'migrations/migration_runner.dart';