+++ lib/types/db_type.dart
@@ -1,1 +1,1 @@
-enum DbType { postgres, mysql }
+enum DbType { postgres, mysql, sqlite }